Consolidate blockdev interface into 3_0
