--- src/dbus_api/blockdev/blockdev_2_0/mod.rs
@@ -1,8 +1,0 @@
-mod api;
-mod methods;
-mod props;
-
-pub use api::{
-    devnode_property, hardware_info_property, initialization_time_property, pool_property,
-    set_userid_method, tier_property, user_info_property, uuid_property,
-};
