--- src/dbus_api/blockdev/blockdev_2_2/mod.rs
@@ -1,4 +1,0 @@
-mod api;
-mod props;
-
-pub use api::physical_path_property;
