+++ src/dbus_api/blockdev/mod.rs
@@ -14,8 +14,7 @@ use crate::{
     engine::{BlockDev, BlockDevTier, DevUuid, StratisUuid},
 };
 
-mod blockdev_2_0;
-mod blockdev_2_2;
+mod blockdev_3_0;
 mod fetch_properties_3_0;
 mod shared;
 
@@ -37,51 +36,16 @@ pub fn create_dbus_blockdev<'a>(
         )
         .introspectable()
         .add(
-            f.interface(consts::BLOCKDEV_INTERFACE_NAME, ())
-                .add_m(blockdev_2_0::set_userid_method(&f))
-                .add_p(blockdev_2_0::devnode_property(&f))
-                .add_p(blockdev_2_0::hardware_info_property(&f))
-                .add_p(blockdev_2_0::initialization_time_property(&f))
-                .add_p(blockdev_2_0::pool_property(&f))
-                .add_p(blockdev_2_0::tier_property(&f))
-                .add_p(blockdev_2_0::user_info_property(&f))
-                .add_p(blockdev_2_0::uuid_property(&f)),
-        )
-        .add(
-            f.interface(consts::BLOCKDEV_INTERFACE_NAME_2_2, ())
-                .add_m(blockdev_2_0::set_userid_method(&f))
-                .add_p(blockdev_2_0::devnode_property(&f))
-                .add_p(blockdev_2_0::hardware_info_property(&f))
-                .add_p(blockdev_2_0::initialization_time_property(&f))
-                .add_p(blockdev_2_0::pool_property(&f))
-                .add_p(blockdev_2_0::tier_property(&f))
-                .add_p(blockdev_2_0::user_info_property(&f))
-                .add_p(blockdev_2_0::uuid_property(&f))
-                .add_p(blockdev_2_2::physical_path_property(&f)),
-        )
-        .add(
-            f.interface(consts::BLOCKDEV_INTERFACE_NAME_2_4, ())
-                .add_m(blockdev_2_0::set_userid_method(&f))
-                .add_p(blockdev_2_0::devnode_property(&f))
-                .add_p(blockdev_2_0::hardware_info_property(&f))
-                .add_p(blockdev_2_0::initialization_time_property(&f))
-                .add_p(blockdev_2_0::pool_property(&f))
-                .add_p(blockdev_2_0::tier_property(&f))
-                .add_p(blockdev_2_0::user_info_property(&f))
-                .add_p(blockdev_2_0::uuid_property(&f))
-                .add_p(blockdev_2_2::physical_path_property(&f)),
-        )
-        .add(
             f.interface(consts::BLOCKDEV_INTERFACE_NAME_3_0, ())
-                .add_m(blockdev_2_0::set_userid_method(&f))
-                .add_p(blockdev_2_0::devnode_property(&f))
-                .add_p(blockdev_2_0::hardware_info_property(&f))
-                .add_p(blockdev_2_0::initialization_time_property(&f))
-                .add_p(blockdev_2_0::pool_property(&f))
-                .add_p(blockdev_2_0::tier_property(&f))
-                .add_p(blockdev_2_0::user_info_property(&f))
-                .add_p(blockdev_2_0::uuid_property(&f))
-                .add_p(blockdev_2_2::physical_path_property(&f)),
+                .add_m(blockdev_3_0::set_userid_method(&f))
+                .add_p(blockdev_3_0::devnode_property(&f))
+                .add_p(blockdev_3_0::hardware_info_property(&f))
+                .add_p(blockdev_3_0::initialization_time_property(&f))
+                .add_p(blockdev_3_0::pool_property(&f))
+                .add_p(blockdev_3_0::tier_property(&f))
+                .add_p(blockdev_3_0::user_info_property(&f))
+                .add_p(blockdev_3_0::uuid_property(&f))
+                .add_p(blockdev_3_0::physical_path_property(&f)),
         )
         .add(
             f.interface(consts::PROPERTY_FETCH_INTERFACE_NAME_3_0, ())
@@ -103,35 +67,6 @@ pub fn get_blockdev_properties(
     dev: &dyn BlockDev,
 ) -> InterfacesAddedThreadSafe {
     initial_properties! {
-        consts::BLOCKDEV_INTERFACE_NAME => {
-            consts::BLOCKDEV_DEVNODE_PROP => shared::blockdev_devnode_prop(dev),
-            consts::BLOCKDEV_HARDWARE_INFO_PROP => shared::blockdev_hardware_info_prop(dev),
-            consts::BLOCKDEV_USER_INFO_PROP => shared::blockdev_user_info_prop(dev),
-            consts::BLOCKDEV_INIT_TIME_PROP => shared::blockdev_init_time_prop(dev),
-            consts::BLOCKDEV_POOL_PROP => parent.clone(),
-            consts::BLOCKDEV_UUID_PROP => uuid_to_string!(dev_uuid),
-            consts::BLOCKDEV_TIER_PROP => shared::blockdev_tier_prop(tier)
-        },
-        consts::BLOCKDEV_INTERFACE_NAME_2_2 => {
-            consts::BLOCKDEV_DEVNODE_PROP => shared::blockdev_devnode_prop(dev),
-            consts::BLOCKDEV_HARDWARE_INFO_PROP => shared::blockdev_hardware_info_prop(dev),
-            consts::BLOCKDEV_USER_INFO_PROP => shared::blockdev_user_info_prop(dev),
-            consts::BLOCKDEV_INIT_TIME_PROP => shared::blockdev_init_time_prop(dev),
-            consts::BLOCKDEV_POOL_PROP => parent.clone(),
-            consts::BLOCKDEV_UUID_PROP => uuid_to_string!(dev_uuid),
-            consts::BLOCKDEV_TIER_PROP => shared::blockdev_tier_prop(tier),
-            consts::BLOCKDEV_PHYSICAL_PATH_PROP => shared::blockdev_physical_path_prop(dev)
-        },
-        consts::BLOCKDEV_INTERFACE_NAME_2_4 => {
-            consts::BLOCKDEV_DEVNODE_PROP => shared::blockdev_devnode_prop(dev),
-            consts::BLOCKDEV_HARDWARE_INFO_PROP => shared::blockdev_hardware_info_prop(dev),
-            consts::BLOCKDEV_USER_INFO_PROP => shared::blockdev_user_info_prop(dev),
-            consts::BLOCKDEV_INIT_TIME_PROP => shared::blockdev_init_time_prop(dev),
-            consts::BLOCKDEV_POOL_PROP => parent.clone(),
-            consts::BLOCKDEV_UUID_PROP => uuid_to_string!(dev_uuid),
-            consts::BLOCKDEV_TIER_PROP => shared::blockdev_tier_prop(tier),
-            consts::BLOCKDEV_PHYSICAL_PATH_PROP => shared::blockdev_physical_path_prop(dev)
-        },
         consts::BLOCKDEV_INTERFACE_NAME_3_0 => {
             consts::BLOCKDEV_DEVNODE_PROP => shared::blockdev_devnode_prop(dev),
             consts::BLOCKDEV_HARDWARE_INFO_PROP => shared::blockdev_hardware_info_prop(dev),
